+++ client/main.tsx
@@ -1,10 +1,11 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App'
+import AuthGate from './AuthGate'
 import './styles.css'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <App />
+    <AuthGate>{(user, signOut) => <App user={user} signOut={signOut} />}</AuthGate>
   </StrictMode>,
 )